fixxed the bug and add back features

+++ src/main/java/dao/WeeklyDAO.java
@@ -10,7 +10,7 @@ import weeks.WeeklyFunctions;
 public interface WeeklyDAO {
 	int createWeeklyDao(WeeklyFunctions weeks) throws DatabaseConnectivityExecption;
 	WeeklyFunctions getWeeklyCalculation(int id) throws DatabaseConnectivityExecption, WeekNotFoundException;
-	double updateWeeklyCalculation(int id, double newWeeklyTime) throws DatabaseConnectivityExecption;
+	int updateWeeklyCalculation(int id, double newWeeklyTime) throws DatabaseConnectivityExecption;
 	boolean deleteWeekbyID(int id) throws DatabaseConnectivityExecption;
 	List<WeeklyFunctions> getAllWeeklyCalculation() throws DatabaseConnectivityExecption;
